third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFilterElement-dom-x-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'x' attribute of the SVGFilterElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS filterElement.getAttribute('x') is "100"
PASS filterElement.getAttribute('x') is "0"
PASS successfullyParsed is true

TEST COMPLETE
